Repository/Service/Controller 나머지 Long→Integer 수정 (2차)
- PriceCalendarService/Impl: companyId 파라미터 Long→Integer
- CompanyService/Impl: getCompanyIdByOfCurrentUser 등 반환 타입 Long→Integer
- DashboardServiceImpl/PaymentServiceImpl: 로컬 변수 Long→Integer
- Repository query 파라미터 Long→Integer (InquiryRepository, PaymentRepository 등)
- Controller 로컬 변수 Long→Integer (InquiryController, PaymentController 등)

+++ src/main/java/com/honeyrest/honeyrest_host/controllerAdmin/InquiryController.java
@@ -31,7 +31,7 @@ public class InquiryController {
                        @RequestParam(defaultValue = "10") int size,
                        Model model) {
 
-        Long companyId = companyService.getCompanyIdByOfCurrentUser();
+        Integer companyId = companyService.getCompanyIdByOfCurrentUser();
 
         Pageable pageable = PageRequest.of(Math.max(page - 1, 0), size);
